Move traits in PMP/internal
